requirements.txt
sarvamai
pandas
streamlit
python-dotenv
google-genai
audio-recorder-streamlit
